Category panel: drop the old panel abstraction
